+++ classes/helpers/currency.php
@@ -44,7 +44,12 @@ class NostoTaggingHelperCurrency
     {
         $id_lang = $context->language->id;
         $id_shop = $context->shop->id;
-        $id_shop_group = $context->shop->id_shop_group;
+        if (isset($context->shop->id_shop_group)) {
+            $id_shop_group = $context->shop->id_shop_group;
+        } else {
+            $id_shop_group = null;
+        }
+
         $base_id_currency = (int)Configuration::get('PS_CURRENCY_DEFAULT', $id_lang, $id_shop_group, $id_shop);
         if ($base_id_currency === 0) {
             $base_id_currency = (int)Configuration::get('PS_CURRENCY_DEFAULT', null, $id_shop_group, $id_shop);
@@ -243,6 +248,48 @@ class NostoTaggingHelperCurrency
                 self::CURRENCY_PRECISION
             )
         );
+    }
 
+    /**
+     * Updates exchange rates for all stores and Nosto accounts
+     *
+     * @throws NostoException
+     * @return void
+     */
+    public function updateExchangeRatesForAllStores()
+    {
+        /** @var NostoTaggingHelperAccount $helper_account */
+        $helper_account = Nosto::helper('nosto_tagging/account');
+        /** @var NostoTaggingHelperContextFactory $context_factory */
+        $context_factory = Nosto::helper('nosto_tagging/context_factory');
+        /** @var NostoTaggingHelperConfig $helper_config*/
+        $helper_config = Nosto::helper('nosto_tagging/config');
+
+        foreach (Shop::getShops() as $shop) {
+            $id_shop = isset($shop['id_shop']) ? (int)$shop['id_shop'] : null;
+            $id_shop_group = isset($shop['id_shop_group']) ? (int)$shop['id_shop_group'] : null;
+            foreach (Language::getLanguages(true, $id_shop) as $language) {
+                $id_lang = (int)$language['id_lang'];
+                $use_multiple_currencies = $helper_config->useMultipleCurrencies($id_lang);
+                if ($use_multiple_currencies) {
+                    $nosto_account = $helper_account->find($id_lang, $id_shop_group, $id_shop);
+                    if (!is_null($nosto_account)) {
+                        $context = $context_factory->forgeContext($id_lang, $id_shop);
+                        if (!$helper_account->updateCurrencyExchangeRates(
+                            $nosto_account,
+                            $context
+                        )
+                        ) {
+                            throw new NostoException(
+                                sprintf(
+                                    'Exchange rate update failed for %s',
+                                    $nosto_account->getName()
+                                )
+                            );
+                        }
+                    }
+                }
+            }
+        }
     }
 }