+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to Semantic Versioning(http://semver.org/).
 ## 2.7.A1
 - Support for Prestashop 1.7
 - Fix non-active product saving
+- Add module version for Nosto calls
 
 ## 2.6.1
 - Change order and product tagging to header / footer